Allow the user to select a Docker instance for the container
