+++ plugins/jupyter-integration/src/main/java/org/polypheny/db/jupyter/JupyterClient.java
@@ -257,93 +257,6 @@ public class JupyterClient {
 
 
     /**
-     * Get contents of any file or directory.
-     *
-     * @param path the location of the file or directory
-     * @param content whether to return inner content ("0" for no content, "1" for return content)
-     * @param format specifies the content format (text, base64, text, null)
-     * @return the HttpResponse from the jupyter server
-     * @throws JupyterServerException if the request fails
-     */
-    public HttpResponse<String> getContents( String path, String content, String format ) throws JupyterServerException {
-        String queryParams = "content=" + content;
-        if ( format != null ) {
-            queryParams += "&format=" + format;
-        }
-        return sendGET( "contents/" + path, queryParams );
-    }
-
-
-    /**
-     * Get file contents encoded in Base64.
-     *
-     * @param path the location of the file
-     * @return the HttpResponse from the jupyter server
-     * @throws JupyterServerException if the request fails
-     */
-    public HttpResponse<String> getFileBase64( String path ) throws JupyterServerException {
-        String queryParams = "content=1&format=base64";
-        return sendGET( "contents/" + path, queryParams );
-    }
-
-
-    /**
-     * Create a new file in the specified path.
-     *
-     * @param parentPath the path to the directory in which the file will be created
-     * @param body serialized JSON object as specified by the jupyter server API
-     * @return the HttpResponse from the jupyter server
-     * @throws JupyterServerException if the request fails
-     */
-    public HttpResponse<String> createFile( String parentPath, String body ) throws JupyterServerException {
-        return sendPOST( "contents/" + parentPath, body );
-    }
-
-
-    /**
-     * Save or upload a file.
-     *
-     * @param filePath the location of the file
-     * @param body serialized JSON object as specified by the jupyter server API
-     * @return the HttpResponse from the jupyter server
-     * @throws JupyterServerException if the request fails
-     */
-    public HttpResponse<String> putFile( String filePath, String body ) throws JupyterServerException {
-        return sendPUT( "contents/" + filePath, body );
-    }
-
-
-    /**
-     * Delete the file or directory in the specified path. A directory must be empty to be able to delete it.
-     *
-     * @param filePath the location of the file or directory to be deleted.
-     * @return the HttpResponse from the jupyter server
-     * @throws JupyterServerException if the request fails
-     */
-    public HttpResponse<String> deleteFile( String filePath ) throws JupyterServerException {
-        HttpResponse<String> response = sendDELETE( "contents/" + filePath );
-        if ( response.statusCode() == 400 ) {
-            throw new JupyterServerException( response.body()
-                    .replace( JupyterPlugin.SERVER_TARGET_PATH + '/', "" ), 400 );
-        }
-        return response;
-    }
-
-
-    /**
-     * Move and / or rename an existing file or directory.
-     *
-     * @param srcFilePath path to the original file or directory
-     * @param body serialized JSON object as specified by the jupyter server API
-     * @return the HttpResponse from the jupyter server
-     * @throws JupyterServerException if the request fails
-     */
-    public HttpResponse<String> moveFile( String srcFilePath, String body ) throws JupyterServerException {
-        return sendPATCH( "contents/" + srcFilePath, body );
-    }
-
-
-    /**
      * Send a GET request to the specified endpoint.
      *
      * @param resource the endpoint the GET request is sent to
