+++ webui/src/main/java/org/polypheny/db/webui/HttpServer.java
@@ -199,9 +199,9 @@ public class HttpServer implements Runnable {
 
 
     private void attachExceptions( Javalin server ) {
-        server.exception( SocketException.class, ( e, ctx ) -> {
-            ctx.status( 400 ).result( "Error: Could not determine IP address." );
-        } );
+        server.exception( SocketException.class, ( e, ctx ) ->
+                ctx.status( 400 ).result( "Error: Could not determine IP address." )
+        );
 
         defaultException( IOException.class, server );
         defaultException( ServletException.class, server );
@@ -213,9 +213,9 @@ public class HttpServer implements Runnable {
 
 
     private void defaultException( Class<? extends Exception> exceptionClass, Javalin server ) {
-        server.exception( exceptionClass, ( e, ctx ) -> {
-            ctx.status( 400 ).json( new Result( e ) );
-        } );
+        server.exception( exceptionClass, ( e, ctx ) ->
+                ctx.status( 400 ).json( new Result( e ) )
+        );
     }
 
 
@@ -391,22 +391,27 @@ public class HttpServer implements Runnable {
 
 
     public void addSerializedRoute( String route, BiConsumer<Context, Crud> action, HandlerType type ) {
+        addSerializedRoute( route, r -> action.accept( r, crud ), type );
+    }
+
+
+    public void addSerializedRoute( String route, Consumer<Context> action, HandlerType type ) {
         log.info( "Added route: {}", route );
         switch ( type ) {
             case GET:
-                server.get( route, r -> action.accept( r, crud ) );
+                server.get( route, action::accept );
                 break;
             case POST:
-                server.post( route, r -> action.accept( r, crud ) );
+                server.post( route, action::accept );
                 break;
             case PUT:
-                server.put( route, r -> action.accept( r, crud ) );
+                server.put( route, action::accept );
                 break;
             case DELETE:
-                server.delete( route, r -> action.accept( r, crud ) );
+                server.delete( route, action::accept );
                 break;
             case PATCH:
-                server.patch( route, r -> action.accept( r, crud ) );
+                server.patch( route, action::accept );
                 break;
         }
     }
